Use a boxed array to dramatically reduce struct size
Summary:
The `Edges` type was using a `[Mutex<...>; 2048]` array for one of
its fields, on the theory that inlining it is a bit more local and
avoids a level of indirection. However, there is evidence that it's
so huge it's blowing the stack in debug mode in a function that's
manipulating one of these.

Instead, push the array into a `Box<[Mutex<...>; 2048]>` so the
containing object is quite small. The level of indirection to find
this object won't have any noticeable perf impact because (1) we
lock the shard we look up, and (2) we do a bunch of work once the
shard is locked. One load won't matter.

Reviewed By: edwinsmith

Differential Revision: D47164630

fbshipit-source-id: b9c493e799d42ad6b9bedd055ebf1693631fefff