Shallow_class_fanout mostly manipulates a DepSet.t instead of an AffectedDeps.t
Summary:
AffectedDeps.t has been stripped off of its fields to the point where it only has two, one of which (`changed`) is only used for the prechecked algorithm.

My first goal is to just manipulate (and accumulate on) a DepSet.t instead of an AffectedDeps.t when computing fanouts.
I'll then likely remove that module altogether.

This is a series of micro refactorings, to make it easier for the reviewer to ensure correctness.

In this diff, Shallow_class_fanout mostly maniputlates a DepSet.t instead of AffectedDeps.t

Differential Revision: D47055097

fbshipit-source-id: e565dede7e3695d03146b7b7a538ca1cee1af20e