Shallow_class_fanout.get_minor_change returns a DepSet.t instead of AffectedDeps.t
Summary:
AffectedDeps.t has been stripped off of its fields to the point where it only has two, one of which (`changed`) is only used for the prechecked algorithm.

My first goal is to just manipulate (and accumulate on) a DepSet.t instead of an AffectedDeps.t when computing fanouts.
I'll then likely remove that module altogether.

This is a series of micro refactorings, to make it easier for the reviewer to ensure correctness.

In this diff, Shallow_class_fanout.get_minor_change maniputlates and returns a DepSet.t instead of AffectedDeps.t

Differential Revision: D47054999

fbshipit-source-id: b6ee2960b84d5b94219bee2f10c624a43987e716