SetPatch::erase bugfix
Summary:
Currently if the Patch has AssignOp, when we erase an element, we will actually insert this element.

The problem is this line

```
assignOr(*data_.remove()).insert(std::forward<U>(val));
```

`assignOr` returns `assign` field if it exists, thus we are actually inserting the element into `assign` field (rather than inserting into `remove` field as intended).

We can fix it by forwarding the call to `remove(...)` that handles multiple elements. This also unifies the implementation of `erase` and `remove`.

Reviewed By: thedavekwon

Differential Revision: D47161489

fbshipit-source-id: a285f49ef10a8896e13a7346714eaa41a9c885d5