Prevent accessing invalidated iterators in TDigest::merge
Summary:
According to the standard, `std::vector::insert` invalidates the end iterator
(and any iterators past the point of insertion). This seems to be benign in
practice but MSVC's iterator debug assertions catch it. We can preserve the spirit
of the original code by replacing the iterators with pointers or indexes. I chose
the latter.

Differential Revision: D47218022

fbshipit-source-id: 96808b20a9483b0ce0a4939edb3de48f48ef4f22