Make member-ops even simpler
Summary:
The HackC Inter conversion had been using the HHVM model of trying to minimize copying by using a base register to store the address of intermediate operations - but this confuses infer.  So instead we are more explicit about COW rewrites.
```
$x[2][3]->y[4][5] = 6;
```
used to produce (inlining the constants for readability):
```
  n2 = $builtins.hack_array_get(&$x, 2, 3)
  n6 = $builtins.hack_array_cow_set(n2.?.y, 4, 5, 6)
```
but now produces:
```
  n2: *HackMixed = load &$x
  n3 = $builtins.hack_array_get(n2, 2, 3)
  n7: *HackMixed = load n3.?.y
  n8 = $builtins.hack_array_cow_set(n7, 4, 5, 6)
  store n3.?.y <- n8: *HackMixed
```
(note the explicit write-back to `n3.?.y`)

Reviewed By: davidpichardie

Differential Revision: D46899984

fbshipit-source-id: 356ab38febd9a8723bee628fb5bffd2a2b54b24c