Make use of NoAutoBound and NoAutoLikes in hhi files
Summary:
Reduce the number of complex attributes in hhi files by making use of `__NoAutoBound` and `__NoAutoLikes` attributes.

This has the effect of allowing `hh_single_type_check --enable-sound-dynamic-type` to be executed without picking up errors in hhi files i.e. to run it in SD mode but without explicit or implicit pessimisation of hhi files. (The presence of `<<__SupportDynamicType>>` on some types in hhi files is necessary only so that our non-implicit-mode tests can make use of builtins such as `vec`.)

Differential Revision: D47089241

fbshipit-source-id: 7ae745d6960f1b378b1bbeecdf46adc448e0b5a4