Loosen invariants around unresolved classes in HHBBC
Summary:
There were invariants that an unresolved class could never be an
"exact" type, only "sub". This was because since the class is
unresolved, you can't truely know if it's exact or not. This ignores,
however, that in some contexts you know it's a-priori, an exact
type. So, when making use of unresolved classes in external workers,
we'd hit the assertion. Remove this invariant, which requires adding
more cases to deal with in the class operators.

We only had an invariant that an intersection list could only ever
contain all unresolved classes or all unresolved classes. This
invariant doesn't actually simplify anything, so remove it.

Reviewed By: arnabde03

Differential Revision: D47304194

fbshipit-source-id: 8a960b1c69f97f932f5e49a09603c576654635ba