Fix handling of include prefix
Summary:
D47195241 changed the order of processing includes breaking one corner case. Fix this by always setting the include prefix if it is not empty.

The whole include prefix manipulation is a mess and we should get rid of it but this is a much bigger change. This is a quick fix that restores the old behavior and unbreaks at least one target.

Differential Revision: D47340496

fbshipit-source-id: 03a26d1e4bd65ef113364d6e511baace4ff2ce75