Fix crash on shutdown
Summary:
The timer_delete() call does not make any guarantees about a potential pending
signal. It is possible for the signal to be delivered during process shutdown
after either the Xenon or RequestInfo structures were already freed.

There is no particularly good reason why Xenon should be using the archaic
signal mechanism. The current usage is incorrect for a number of other reasons,
such as obtaining a lock in a signal handler.

This diff replaces it with a background thread that generates Xenon events at
the configured interval. This thread is properly exited during process shut
down.

Reviewed By: ricklavoie

Differential Revision: D47330040

fbshipit-source-id: baa028a5662665e8bf9432dde703ac1d7c61fe05