Fix hack_macros position
Summary:
The rust linter showed that the hack_macro was unnecessarily cloning in a few places. It turns out that although I intended to copy the position once and then use that there were a few places where I used the original TokenStream directly.

i.e. code written like:
```
hack_expr!(pos=X, ...)
```
was supposed to turn into:
```
{
  let __hygienic_pos = X;
  [MyExpr(__hygienic_pos.clone()), MyExpr(__hygienic_pos.clone())]
}
```
but instead was:
```
{
  let __hygienic_pos = X;
  [MyExpr(X.clone()), MyExpr(X.clone())]
}
```
which is problematic if `X` is a complex or expensive expression.

The changes are a little more complex just so I could create the identifier for `__hygienic_pos` and pass it around as a token as opposed to relying on it being well-named.

Reviewed By: mjhostet

Differential Revision: D47297457

fbshipit-source-id: 5994dd3ebed5ef0ba20fcf63f07ad273fbfceffb