Ban Case Types from Being Used as Enum Base Types
Summary:
Runtime enforcement for case types is being worked on. As part of this oulgen suggested banning case types being used as base types for enum and enum classes due to some inconsistent behavior he has uncovered (listed in attached task).

There are still other more core/common aspects of case types we are prioritizing working on, so for now we will make it a type error to use case types in these parts of the program. When we either:
1. Get requests to use case types in these positions
2. Audit more thoroughly the runtime behavior and expected behavior for enum/enum class enforcement

we can remove this check.

Reviewed By: oulgen

Differential Revision: D47351127

fbshipit-source-id: 6a2a1e94b3235b82a30671e9f2d64e9330556eff