python: add unit tests for current behavior of double leading underscores in field name
Summary:
NOTE: this diff is to test the **current** behavior, not the **desired** behavior.

Context: in Thrift, field name that begins with two or more underscore characters and does not end in two or more underscores is legal.

But there's also Python's [private name mangling](https://docs.python.org/3.11/reference/expressions.html#atom-identifiers) rule.

Consider following example:

```
struct Reserved {
  8: string __mangled_str;
  9: i64 __mangled_int;
}
```

We have two design choices:

1. Do NO mangling:
That's how non-primitive-type fields work today, as shown in unit tests added in this diff.
But primitive-type fields with such names (e.g. `i64 __mangled_int`) will straightly fail to compile. That's because we use `__slots__` to optimize performance of those fields, and in such case name mangling happens in compile time and there's no way to avoid it.
2. Always mangle:
That will change the current behavior of those non-primitive-type fields with such name.
But it will allow primitive-type fields with such name to work.
More importantly it allows us to optimize non-primitive-type fields with `__slots__` in the future, which is what I am working on now.

I propose we go with 2. (always mangling), see next diff for implementation.

Reviewed By: Filip-F

Differential Revision: D47414207

fbshipit-source-id: a890567af87f4da660a67f0834038489dc2afba3