Add enforcement for typed locals
Summary: Enforce typed locals. Also add naming errors when a declaration shadows an already in scope local, and when a declaration control-flow joins another declaration (or assignment). This is not yet complete over all of the kinds of statements and lambda expressions. TODO comments indicate what still needs implmenting in subsequent diffs.

Reviewed By: andrewjkennedy

Differential Revision: D47092623

fbshipit-source-id: 0f1250a4796cd86cff48d3256c35e267d0071a47