Fix F14MapFallback `insert_or_assign` behaviour
Summary:
When `F14MapFallback` is used and we call `insert_or_assign`, the call
chain is the following:
1. `F14::insert_or_assign`,
2. `F14::try_emplace`,
3. `F14::emplace`,
4. `std::unordered_map::emplace`.

The problem is: `std::unordered_map::emplace` will create a `__node` from `__args`
no matter if we actually inserted an object or not. See implementation [here][1].

This means, by the time we try to assign inside `F14::insert_or_assign` the
object might be already in the «moved from» state.

Reviewed By: dmm-fb

Differential Revision: D47441603

fbshipit-source-id: bd2e3d17445b8c3c0124c36d78c0f9ee575bc17b