fix arity of __self__ static call in traits
Summary:
Right now there is an asymetry that make Pulse reasoning quite complex. A
static method call on "normal" types is exported "as is": `C::f($x)` is
exported as `C$static.f($x)`. However for static calls on self and parent, we
inserted a new argument, `self`. `self::f($x) -> __self__$static.f($x, self)`

Let's not do that, it's the job of Pulse to do this insertion, and this
difference makes Pulse's job quite hard.

Reviewed By: aorenste

Differential Revision: D47514522

fbshipit-source-id: 0fbdd984d6cf53431303c0ad10149fb3ac923d32