Wrap arraydata printer in error handling mechanism
Summary:
When there are errors in pretty printing a type (e.g. when there are missing symbols so look ups fail), builtin commands like `bt` fail because they implicitly use the pretty printers for displaying each frames variables.

We already wrap the basic pretty printers in an error catching mechanism; this wraps the synthetic child provider class for arrays' `update()` method in an error handler as well. I was seeing failures using lldb on production cores when printing the backtrace.

Reviewed By: jeffreytan81

Differential Revision: D47490134

fbshipit-source-id: 2f91c3e209038a072c9f8b97ecd3364085fef5d2