Allow wildcards in lambda type hints
Summary: We'd like to support wildcards in hints on lambdas in the Hack typechecker e.g. writing `(vec<_> $x) ==> ...` or `($x) : vec<_> ==> ...`. Before we do that, there is some unfinished business in HackC dealing with the new `Hwildcard` type hint AST node. Essentially it should treat it as "any type" as far as enforcement goes, not really any different from `mixed`.

Differential Revision: D47437871

fbshipit-source-id: 0e1dfad5663c3ae5acefdad66e48a04c8a4326c1