Keep UNSAFE_CAST in Tast
Summary:
Instead of replacing with `Upcast`, keep the expression and just rely on the function call checking against the `hhi` signature:
```
function UNSAFE_CAST<<<__Explicit>> Tin, <<__Explicit>> Tout>(
  Tin $t,
  ?\HH\FormatString<nothing> $msg = null,
)[]: Tout;
```
Now under pessimisation, there will be a `supportdyn<mixed>` bound on both type parameters. So we will be checking that both source and target types support dynamic. We will be checking that the argument expression is of type `Tin` as specified. And we can cast to anything safely, as long as it supports dynamic, and the result will end up with a like type (due to pessimisation).

Differential Revision: D46001210

fbshipit-source-id: f38ad5a50b0f2669acea3c7e3889d71b0d48521e