Use Twildcard for missing types on lambda parameter and return types
Summary:
Instead of abusing the `Tany` type to represent omitted type hints for lambda parameters and returns, use the new `Twildcard` representation. This gets automatically localized to a fresh type variable.

Note that actual wildcards in lambda hints will get the same treatment, but for now these are rejected at the naming phase.

Differential Revision: D47547078

fbshipit-source-id: aa45ce93306ad5c7edae58d7e7d63dd40ace2cee