Add support to throw thrift exceptions
Summary:
These changes are needed in order for SR failovers to be triggered for internal carbon result codes. These result codes are mapped to thrift exceptions, which SR can be configured to catch and act on. Note that SR cannot see the actual carbon result, as this is a serialized field in the reply.

Note that the exceptions are defined in the IDL and each reply that can throw is annotated.

Reviewed By: disylh

Differential Revision: D43925205

fbshipit-source-id: e77bfbe0315ff6b5262bb2bfbd88886d786c21c3