Add a program member to t_named
Summary:
If it has a user-defined name that name must appear in some thrift file, which means the associated program can be made available. This helps with handling definitions generically.
Some definitions that do not currently track program (e.g. enum_value) are left with null programs because it is not likely to be useful, even though we could pass it through if we wanted.

Reviewed By: vitaut

Differential Revision: D47560058

fbshipit-source-id: 7513b79afbb4b87bb2454bb82b7b755407d77320