Simplify Case Types in Intersections
Summary:
# Context
Hiding that case types are unions leads to suboptimal types when computing intersections. For instance consider the following case type:

```
case type CT = int | vec<int> | shape('val' => int, ...);
```
The following intersections will fail to simplify:
```
1. CT & vec<mixed>
2. CT & shape(...)
3. CT & not int & not vec
```
Given what we know about data types it should be possible to simplify these intersections into the following types respectively
```
1. vec<int>
2. shape('val' => int, ...)
3. shape('val' => int, ...)
```

Of particular interest are examples 2. and 3. Example 2 will arise from performing the following type assertion `$ct as shape(...)`. If simplification can occur then the field `val` can be safely accessed. Example 3 arises from a series of type tests like:
```
if ($ct is int) {
  ...
} else if ($ct is vec<_>) {
  ...
} else {
  $ct['val']; // Error
}
```
In the final else branch example 3 arises and if it is not simplified will result in a type error.

This diff aims to allow such simplifications to occur

# Solution

The solution implied here is to utilize `Typing_case_types.filter_variants_using_datatype` that was introduced in {D46872793}. By collecting all other types present in the intersection as an intersection type, we can use this as a filter on the variants for a case type.

If after filtering the resulting type is not a union type, then it is clearly simpler than using the case type directly so we substitute that simplified form into the intersection and attempt simplifying the full intersection type again. If the type is still a union after filtering that indicates more than one variant is still potentially available so we leave the case type as the final result in the union.

This needs to be done in a few places within `Typing_intersection`, namely `Typing_intersection.intersect` and `Typing_intersection.intersect_lists`. We also need to be careful to call `Typing_intersection.intersect` again, but only if a simplification occurred. If we recurse into `intersect` without checking for simplification it can lead to infinite recursion.

I'll walk through how this diff helps us to arrive at the simplification for each of the three examples
## Example 1
```
CT & vec<mixed>
=> vec<int> & vec<mixed>
=> vec<int>
```

## Example 2
```
CT & shape(...)
=> shape('val' => int, ...) & shape(...)
=> shape('val' => int, ...)
```

## Example 3
```
CT & not int & not vec
=> (shape('val' => int, ...) & not int) & not vec
=> shape('val' => int, ...) & not int
=> shape('val' => int, ...)
```

Reviewed By: vassilmladenov

Differential Revision: D47388802

fbshipit-source-id: da35b0f22a3105faa4fc8069c4bebc6fd916bca9