Abstract constant support
Summary:
This was a 'todo' for abstract class constants so I could think about how to handle them.

We read/write class constants by either loading the singleton (for a reference like `C::X`) or loading the static 'this' (for a reference like `self::X` or `$a::X`).  Since that's a virtual field access it should work for abstract class constants as well.

So to handle abstract class constants we just ignore them during initialization with the expectation that they'll be initialized by a subclass.

Reviewed By: davidpichardie

Differential Revision: D47594138

fbshipit-source-id: 44525ac7f5765286f18659a64352b28c4b4328a4