TypeIntersectionConstraint
Summary:
Currently there are a number of places where we have a vector of TypeConstraint which has the meaning of an intersection.

For a declaration of:
```
class C<T as A as B>
```
`T` would have an upper bound of `[A, B]` which means that `T` must be both a subtype of `A` and a subtype of `B`.

For case types we're adding unions which will make it confusing in the code whether a vec is intended for use as an intersection or union.

This change adds `TypeIntersectionConstraint` which is a representation of a vector of `TypeConstraint` with the intended use as an intersection.

|`CompactVector<TypeConstraint>`=>`TypeIntersectionConstraint`
|`VMCompactVector<TypeConstraint>`=>`VMTypeIntersectionConstraint`
|`TinyVector<TypeConstraint>`=>`TinyTypeIntersectionConstraint`

Reviewed By: oulgen

Differential Revision: D47245752

fbshipit-source-id: 389923904ab648a2f284133fcca5524f5be45170