capi supports union
Summary: This is similar to the work done with struct: extract the tag and cpp value from the internal data tuple for cpp2 from python; construct the internal data tuple for cpp2 to python.  The union tuple is always size 2, where the first value is the int tag and the second value is the set field.

Reviewed By: Filip-F

Differential Revision: D47420800

fbshipit-source-id: 96fb166cd64d11b1f2e435ee954a5884d4495368