don't pass in globalrev
Summary:
Our SymbolIndex API asks us to pass in a globalrev (and has a costly 1-2s hg query to find one if we don't pass it in). However, it passes it to glean API which uses hashes instead. That doesn't make sense.

For now, I'm just deleting all mentions of globalrev. My design is that the ocaml side should always simply expect that it's getting the "default" glean repo hash, whatever it is. (hopefully the latest).

Reviewed By: mheiber

Differential Revision: D47668544

fbshipit-source-id: cfb3c675c2639d0579caf1ae02971b179e3e3e6c