hh_single_complete support
Summary:
This diff adds glean support to hh_single_complete. This way we'll be able to write end-to-end tests against glean.
```
hh_single_complete --glean-repo-name www.autocomplete --auto-complete ~/c.php
```

It also, as an incidental cleanup, makes symbol-index (for top-level symbol autocomplete) be aware of HHI builtins as well as the files provided.

Here's a detailed list of what's in the diff:
1. It used to call  `SymbolIndex.initialize ~provider_name:"LocalIndex"`. I changed it to `~provider_name:"CustomIndex"` (which is Glean)
2. Previously this initialize call was done while parsing command-line arguments, which didn't feel right. I moved it later to the same place where the naming table is set up. This has the additional benefit that EventLogger has been initialized (necessary because CustomIndex initializer writes to HackEventLogger)
3. The behavior of glean and autocomplete is pretty heavily dependent on auto-namespace-map, e.g. `--auto-namespace-map '{"Vec": "FlibSL\Vec"}' `. It's only because of this that autocomplete knows that Vec\chunk should become a glean lookup for FlibSL\Vec\chunk. Previously auto-namespace-map was being read as a command-line argument but then IGNORED. I changed it to respect the command-line argument. (If the user provides --root then it instead gets read from .hhconfig).
4. Similarly, the --no-builtins command line argument was being ignored. I changed it to be respected.
5. The code used to call `scan_files_for_symbol_index` to add the symbols from the current file-under-test into the autocomplete top-level-symbol delta. I inlined+moved this function into a more natural place, just after SymbolIndex.initialize

The code used to build the symbol delta only for the current file-under-test. I changed it to also build the symbol delta for HHI files (if included), so as to be the same as how the naming table delta was being calculated. I figured this made more sense, was more self-consistent. However, I had to add the flag "--no-builtins" to some tests so as to get the output they were expecting.

Reviewed By: mheiber, vsiles

Differential Revision: D47678867

fbshipit-source-id: f6354852fce44b0ee7efd4ff734ca17d696f1494