delete glean go-to-impl experiment
Summary:
In Nov-Dec 2019, intern ygoldin worked on a prototype to power go-to-impl from a custom glean database. Diffs: https://fburl.com/diffs/3z0ukq4e

We didn't use that implementation, but are now working on a new implementation for find-refs that's backed by the main hack glean index.

This diff deletes Yael's prototype.

I'm doing this now because the code gets in the way of all the glean changes I want to make.

Reviewed By: mheiber

Differential Revision: D47683056

fbshipit-source-id: b79b010ad47f66deafb657c68e4bdde0c38e8072