Tolerate supportdyn<null> in nullsafe method invocation
Summary:
Sometimes we end up with `supportdyn<null>` in inferred types, even though the `supportdyn` here is redundant. This comes about because `supportdyn<#0>` is created and then type variable `#0` is later resolved.

Method invocation should be resilient in the presence of such null types.

The test is contrived, involving explicit use of `supportdyn`. But this has occurred in the wild under implicit pessimisation.

Reviewed By: madgen

Differential Revision: D47751375

fbshipit-source-id: ec81fe5f2633656fb928d2fc1f1742942561060e