Point to correct field definition for optional shapes
Summary:
Shape subtyping creates fake fields and might use the wrong reason as we fold over a set of all fields across both shape types involved in subtyping.

This makes it confusing for users to find the source of the error.

This diff fixes the problem by pointing not to the field (which might come from either type), but to the reason for the type having the optional field.

Reviewed By: vsiles

Differential Revision: D47753751

fbshipit-source-id: e7271ff54fb22492bc1f160a6984161473f0dff6