Fix truncated types in "Extract shape type"
Summary:
See snapshot test.

**Design choice**

Magic number 2048, see the comment in code

I chose not to auto-format the extracted shape type because doing so for large shape types would fill up the user's screen and make it hard to see what the refactor did. (And also I'm prioritizing other work-in-progress).

We can revisit formatting in future, for now users can format-on-save or `arc lint -a` or use cmd-p-format

**Potential follow-on work**
- Potential future work not taking on right now, will do if users ask for it: handle giant types in return and parameter position for the "Extract into method" refactoring/code action: T158766159

Reviewed By: madgen

Differential Revision: D47589425

fbshipit-source-id: 2deb75de63c750a0823231f082b45552324944db