remove memcpy when serializing header
Summary:
Changes the header so that it doesn't use memcpy to write to the byte array that is being passed in. Changes the 24-bit encoding to use bit-shifting to write directly to backing byte array.
Checks to see if you its possible to write before serializing data to the header.

Also the DCHECK was incorrect. The frame length is supposed to be unsigned so it correct thing to do would be to make it less than or equal to 1<< 24

Reviewed By: sazonovkirill

Differential Revision: D47825054

fbshipit-source-id: b4ee13dcf9fc93937601bfdc92da80f0434b92af