Revert D47775177: Multisect successfully blamed "D47775177: [curl-8.1.2] buckify_tp2" for test or build failures
Summary:
This diff is reverting D47775177
D47775177: [curl-8.1.2] buckify_tp2 by yns88 has been identified to be causing the following test or build failures:

Tests affected:
- [deeplearning/projects/fairstore:download_helper_tests - test_good_download (deeplearning.projects.fairstore.tests.download_helper_tests.TestSomething)](https://www.internalfb.com/intern/test/562950026458083/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2583198
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: yns88

Differential Revision: D47825013

fbshipit-source-id: 33018b0d3d2cf726ced5853ca1abf44f7cad6c7d