Create type definition for Tshape
Summary: This is in preparation for the follow up diff which converts this tuple to a record. The noteworthy change here is that the use of the alias changes the compiler representation from a constructor with 3 elements to a constructor with a single 3-tuple element; this is reflected in the Rust changes.

Reviewed By: rjbailey

Differential Revision: D47742173

fbshipit-source-id: 820563eb0469359d16a568b946130e560c37a356