Don't Require Reified Generics as Part of a Reified Type to be Enforceable
Summary:
# Context
We allow a subset of reified types to be used as part of `is`/`as` expressions. Currently the type checker requires that a generic that is used as part of one of these expressions must be reified **and** enforceable. This is too restrictive for instances when the generic appears as part of a reified type used to instantiate a reified class generic. For instance:
```
class C<reify T> {}

function f<reify T>(): void {
  '' as T; // Error `T` must be enforceable
  '' as C<T>; // Ok `T` is used as part of a reified type that instantiates a reified generic
}
```

This is ok because the reified type is used to construct a type structure that will be compared against another type structure, and not an actual value. The diff aims to loosen this restriction.

# Mitigation
The runtime already properly handles this case so the only change needed is in the type checker. In this case we need to update `Typing_enforceable_hint`. Inside `on_class` when we are checking a type argument that is a reified generic we set a flag in the environment that we are currently inside a reified class generic position. If this is the case we don't emit errors for unenforceable reified generics that we encounter.

While fixing this I also noticed another bug where we were allowing wildcard types to appear as part of a reified generic. This feels incorrect because wildcards are not valid types so they should not be used as a type witness.

Reviewed By: vassilmladenov

Differential Revision: D47488190

fbshipit-source-id: 489dd1121fd8701f26828616556c3626cefdbe2c