Fix FastStaticBool for ABIs with unsigned char
Summary:
`char` is unsigned on ARM [1][2]. Even if we'd stored -1 to indicate
false, integer promotion would kick in for `f > 0`, and we'd end up
zero-extending to 255 instead of sign-extending to -1, and thus
erroneously return true. See https://godbolt.org/z/oaT3964MG for another
demonstration. Fix this by explicitly using `signed char` instead.

This seems like a general portability footgun between x86 and ARM.
Unfortunately, `-Wsign-conversion` seems to be the only thing that would
catch this particular instance, but it might have lots of false
positives as well (e.g. where `-1` is used as a sentinel value and we
directly compare against it), so I'm not sure of a good general strategy
to prevent this.

[1] https://github.com/ARM-software/abi-aa/blob/main/aapcs32/aapcs32.rst#arithmetic-types
[2] https://github.com/ARM-software/abi-aa/blob/main/aapcs64/aapcs64.rst#arithmetic-types

Reviewed By: ispeters

Differential Revision: D47850892

fbshipit-source-id: cb093d75ce86f8cefd21aaac76668fcbd30d2487