Properly Handle Negated Special Classes for Case Types
Summary:
# Context
This is a follow up to D47388802 to fix the temporary hack of instantiating class types that might have generic parameters without their generics when computing the set of data types associated with a class.

While cleaning up this code I also discovered a subtle bug that led to incorrectly concluding that `vec` and `KeyedContainer` do not have overlapping data types, even though `vec`s are `KeyedContainer`.

# Mitigation
I decided to centralize the logic for computing data types from a class into its own function that operated only on the class name and not the type arguments. For the special interfaces we know how many type arguments they expect, so I handled them manually by instantiating them with fresh type variables.

Since I do not care if the generics matched, I needed to switch from `is_subtype` to `can_subtype` so that propositions involving type variables are considered satisfied. I tried testing some of the other existing variations of subtype already in `Typing_utils` but none of them behaved the way I wanted.

An alternative would be to instantiate the generics with the any type instead of fresh type variables, but I figured that would be less desirable given our plans to eventually get rid of Tany.

Reviewed By: vassilmladenov

Differential Revision: D47817635

fbshipit-source-id: 7570a02f4e0d5dc40379d169529b7f4ca71cbace