support switch to v2
Summary:
This diff's goal is to allow a JK to rollout the change to glean v2 autocomplete.

## Prior to this diff

(1) hh.conf has these (which can be overridden with --config)
```
symbolindex_search_provider = NoIndex  # hh_server doesn't search
ide_symbolindex_search_provider = CustomIndex # ide uses glean
```

(2) .hhconfig is allowed to have `glean_reponame = ...` but currently doesn't. Instead it defaults to "www.autocomplete". It can also be overridden with --config.

(3) The glean logic is hard-coded to always use C++ bindings, JSON queries, hack.1 schema.

## After this diff

(1) There's a new hh.conf setting, which can be overridden by --config and JK "hack/config:glean_v2":
```
glean_v2 = false  # should we use glean v2
```
If it's true, then in ServerConfig.load we disregard the glean_reponame we got (from .hhconfig or from default) and replace it with "hack.www.light". *During this transition period, we won't be able to configure which DB to use for v1, and the only way to configure the DB for v2 will be with --config*.

(2) The glean logic switches on the reponame. If it is set to "www.autocomplete" then it will use the old C++ bindings, JSON queries, hack.1 schema. If it is set to anything else then it will use new Rust bindings, Angle queries. hack.6 schema.

Why did I chose to switch on reponame rather than anything else? It's because reponame is already deeply embedded, and passed through everywhere it's needed, and indeed it already implies v1 vs v2. In any case the GleanOptions.t code is bad (shouldn't exist) and I didn't want to extend it.

(3) Rollout will proceed first by `hh lspswitch --config glean_v2=true` and testing locally, then by rolling out the JK to 50%. The value of the glean_v2 flag is stored in scuba in the `rollout_flags` column so we'll be able to do A/B comparisons.

## Future plan: after rollout has complete

(1) I will change ServerConfig default from "www.autocomplete" to "hack.www.light" (or "hack.www.incr" depending on what we decide/measure)

(2) I will remove the glean_v2 flag, and I will remove the old logic/hack.1 code. Instead all code will go through rust/hack.6

(3) The choice of which glean DB to use will therefore reflect what's in .hhconfig or in ServerConfig default

## Other changes in this diff

The old glean code, if it got a failure in response to a query, would attempt to reconnect to a new glean server just once. However this reconnect logic had a bug where it disregarded the setting of glean_reponame and instead was hard coded to use the default. That won't do, especially now that we're basing decisions on the reponame. I therefore fixed this. I fixed it by storing glean_reponame into sienv, "search index environment", and passing it to all functions which might need to reconnect.

The old code had two layers, `glean.ml` which attempted to be a schema-neutral API for accessing glean via json, and `glean_autocomplete.ml` which was specifically for hack.1 autocomplete schema. But we've learnt that the Rust API for glean is strongly typed, so this division no longer makes sense. I accordingly combined the two.

Reviewed By: mheiber

Differential Revision: D47703370

fbshipit-source-id: 670c53177618994a53beec19102dde16a703e0d0