factor out SearchTypes from SearchUtils
Summary:
This diff creates the following layering of modules, in order:
1. SymbolIndex -- top-level module
2. CustomSearchService -- this provides searching via glean
3. SearchUtils -- contains "si_env", the environment in which is stored the current configuration of the search service (including the current configuration of glean)
4. Glean -- contains the FFI bridge to glean
5. SearchTypes -- contains primitive interop types, e.g. SI_Kind, needed by the glean FFI

Why?

(1) Prior to this diff, SearchUtils and SearchTypes were combined. This meant it was impossible for the Glean FFI to mention primitive interop types. That was fine using the old C++/JSON glean bindings, where Glean simply returned an uninterpreted JSON blob. But now we've moved over to Rust/Angle glean bindings, we get back from glean a strongly-typed record, which we want to send over FFI as strongly-typed SI_Kind.

(2) Prior to this diff, there was way to store within si_env a "handle" to the current glean connection. It didn't belong in SearchTypes (since it's not a type). Therefore, the handle had previously been stored in a global variable. This diff allows it to be defined in the Glean module, where it belongs.

(3) I also moved SearchUtils.is_valid_* into SearchTypes, because I want those functions to be re-used by the glean module for now.

There's nothing interesting to review in this diff. It's all just tedious search/replace.

Reviewed By: mheiber

Differential Revision: D47793779

fbshipit-source-id: 2488e12e01e5f582492ecd08399265bfc1aacf5d