make customSearchService non-global
Summary:
The current glean interop stores a "handle" to the current glean connection in a global mutable variable.

I never like those! And in Rust, they're more trouble than they're worth.

In this diff I change the handle to be a member of si_env, which already stores everything else about what the symbol index needs.

------------
> Note from mheiber: CI has identified that this change causes a regression in binary size. However, none of the binaries are things used in prod, they're only used in testing or in one-off tasks by our time. My two cents is that this is OK. (even assuming the signal is accurate)

ljw1004: I think these binary size regressions are part of the horrible pact we've made with FB infrastructure. Previously, the handle was stored as a global variable hence bypassed normal build dependency chains. Now that the handle is stored properly, we suffer build dependency, but I don't think there are any ocaml solutions that are clean enough to be worth using. (In OO we'd do some kind of dependency injection but that's just not worth it here).

As Max noted, the regressions are only in tests. Our production binaries already suffer the binary size regression...

Reviewed By: mheiber

Differential Revision: D47796938

fbshipit-source-id: 69fff1f9ef460ed33830cba2252ebba0259ca639