deprecated wrapper filename is determined by definition-position
Summary:
As explained in the previous diff, I'm on a mission that the "rename action" which is used for many things, including marshalling from hh_client to hh_server, should not include Relative_path.t (since that can't be marshalled).

Prior to this diff: the ServerRename.get_deprecated_wrapper_patch took both `filename` and `definition` to specify where the wrapper should go. It got `filename` from the rename-action-structure. *But what would happen if the filename+definition disagreed?* It wasn't a good API.

This diff changes it so solely `definition` is used to decide where the wrapper should go.

Reviewed By: nt591

Differential Revision: D47898287

fbshipit-source-id: 43095148fa63ed2bd0b06d5a97a6c72cc185a892