delete old shell-outs
Summary:
This diff deletes some command-line arguments
* `hh --ide-find-refs` and ServerRpc.IDE_FIND_REFS (but leaves `hh --ide-find-refs-by-symbol` and IDE_FIND_REFS_BY_SYMBOL)
* `hh --ide-go-to-impl` and ServerRpc.IDE_GO_TO_IMPL (but leaves `hh --ide-go-to-impl-by-symbol` and IDE_GO_TO_IMPL_BY_SYMBOL)
* `hh --ide-refactor` and ServerRpc.IDE_REFACTOR (but leaves `hh --ide-rename-by-symbol` and IDE_RENAME_BY_SYMBOL)

It used to be that Nuclide language server would shell out to these "hh" commands. Five years ago we changed the language server to have a persistent connection to hh_server so it no longer shelled out to the command-line, but instead did the Rpc call directly. Then in June 2023 we switched over to `ide_standalone` which shells out to `hh --...-by-symbol`.

This diff also deletes
* `hh --ide-highlight-refs` and ServerRpc.IDE_HIGHLIGHT_REFS (but leaves ServerHighlightRefs.ml, which is invoked directly by clientIdeDaemon)

The same telemetry as above shows that no one has used this command for 90 days. Again it used to be part of Nuclide and has been replaced by serverless_ide.

Reviewed By: nt591

Differential Revision: D47897122

fbshipit-source-id: cc631c665dc6d6372c63848985f8b8c3fd297b9c